Add Storm ORM
Add support for running the storm ORM, it is disabled by default and
can be enabled by setting STOQLIB_USE_STORM=1.
Most of the tests passes, there are a few regressions and most applications
are not tested yet.

It includes SQLObject compatibility layer from Storm upstream which has
been forked and modified.