Enable lazyUpdate for AccountTransaction
lazyUpdate in sqlmeta means that the object must be sync:ed or referenced
by another table (which isn't lazyUpdate:d) before an INSERT/UPDATE
query is executed.

This means that we have to call sync() in quite a few places where we
want the items to be in the database.