Add a new api to help controlling widgets sensitivity on apps. r=romaia
That new api was made to workaround a problem with some apps where we have an
open inventory, that would disable some widgets but, other events made them
available.
Now, we add them to a sensitive_group and, when we call self.set_sensitive for
app widgets, if they belong to a sensitive_group, all of it's validation
function will be consulted and, if any returns False, the widget won't be set
sensitive.

As some TODO comments points, we should change and only use
'self.set_sensitive' instead of 'widget.set_sensitive' on apps, as the first
take sensitive groups in consideration and that's a good defensive way to know
we won't have widgets with wrong sensitivity set.

Also, corrected a problem where inventory infobar would appear on screen as
much as the app were activated (e.g. change application).