Warn the user about negative balance when closing Till instead of crashing.
This is a workaround to a problem that should not exist, but happened for some
user. The gui should block all kind of interactions that could lead to
'close_till' being called. If that happens again, the user will see a warning
message (instead of a crash) telling what the problem was, and then they can
solve the issue by themselves.