Add an api for adding deleted objects on StoqlibTransaction.
This was done mainly to prevent a bug where deleted objects could be on both
modified_objs and created_objs, calling their hooks and maybe crashing because
ORMObject made them obsolete.

Also, it's a good way to emit events for objs deletion, and can be a lot useful
(together with creation/deletion) when writing plugins.