Don't try to convert column names to Mixed as this corrupts joinColumn on Joins.
Actually, that method is only used for joins (even it's name say it's used for
all db columns). The problem was hidden as we always did joins with only the
id underscore (e.g. 'sellable_id'). Trying a joinColumn with something like
'base_sellable_id' would result in 'baseSellableID' instead of
'base_sellableID', and sqlobject does not now that column by that name (the
second form is the one used by it).