#4051: Busca com acentuação
This will make queries a bit slower, as we need to convert all the table
rows to the normalized form when searching. That can either be avoided by
- manually creating indexes for these situations
- switching over to fulltext searching.

For large databases with > 100'000 items fulltext searching is a requirement
anyway and this makes the user experience better with smaller databases.

This was tested with a database with 2300 items, it changed the search
time from ~16ms to ~40ms.