Set status before, or we will have problems with invoices created on Stoq.
On update_remote phase, if the invoice isn't marked as paid, we will try to
capture it online, that is, we will ask for the gateway (paypal, pagseguro,
etc) if the payment was already paid. Lets make sure that if we paid it here,
we won't try to do that.