Differentiate between Sale/Purchase payments and their Group payments.
The group holds all payments, even those cancelled. On sales and purchases, we
are only interested on valid payments (the ones that weren't cancelled).
That is, when printing an invoice, checking if the sale/purchase was paid, etc,
we are only interested in those valid payments.
On the other hand, when viewing sale/purchase details, we want to get the whole
payment history.
So, sale.payments and purchase.payments will show those valid payments starting
from now. If one needs to get all the payments, including the invalid ones,
they should use sale.group.payments and purchase.group.payments.

r=jdahlin