Remove comments from ship/invoice and don't send email about invoice to client.
Comments are pointless right now. All needed information is set on Magento.
Also, do not send invoice emails to clientes. When they are created, they are
not necessarily captured. So, one can receive and email saying that he paid the
sale, but actually, he still has the bill on it's hand.