Hopefully solve all problems with payment gateways.
The workflow of magento with gateways is a little hard to understand. When the
gateway receives the payment, it should tell magento to create an invoice and
mark the sale as paid. If the user did not click on 'Confirm the payment on
store page', there's no way for magento to know it the payment was paid or even
if the user cancelled the payment on gateway. On that case, the stoq user needs
to go to the gateway page and manually set each payment that was paid to send a
message to magento. That way, we will bring it to Stoq, and set the sale as
paid on here.