Do not limit results on combos.
Limiting results on combos can lead to errors and does not make sense. On
searchs, it's ok to limit because the user can filter it and find what he is
looking for. On combos, for example, if limiting for example a list of clientes
in a combo, some clients with names starting with a letter of the end of
alphabet won't apear on the list, and there's no way to get them there other
than modifying the parameter MAX_SEARCH_RESULTS.

r=romaia