Just create Magento objects if they don't really exist.
Check when creating a magento object by events if they already exist. We are
going to add tabs on some editors and those objects will be created there
(before the event emission, since they are emitted on transaction.commit).
We cannot remove those events because those objs can be created outside the
editor (in an import for example).
Also, correct the way we create images to support services too.