Fix orm_startup not being called on shell bootstrap.
Missed that on my last patch. Since we are not registering a station or even
checking for the schema on setup, we need to call orm_startup (it would be
called in any of those situations). Maybe that whole bootstrap and setup could
be changed to remove a lot of duplications.