Fix sintegradialog date filter populate.
Since months are list based, January for example is the position '0' and not
'1'. So, it needs to get the month at the index start.month - 1 instead of
start.month directly.
Also, make the name more translation friendly. There is no reason to translate
a '%s %s' string if the '%s's can't change position.