Do not close the dialog when BaseEditor.confirm doesn't really confirm.
In some places, validate_confirm will return False and them the dialog should
not get confirmed. Check for that return value on dialog and avoid closing it
even wherever the 'confirm' event emission returns.

Add a workaround for dialog too because, if one used BasicDialog (or a
subclass) directly, he probably won't connect to the event and therefore, it'll
return False and the dialog won't ever get closed. We should find a better way
of handling that in the future.