Add an option to show the changelog when Stoq is upgraded.
When Stoq is upgraded, a message bar will be show to welcome the user to
the new version and present a button that he can click to display the
changelog. That will probably solve a lot of problems that some feature
changes and the users don't know where they are now.

r=romaia