Pass safe='' to urllib.quote for magento image's name quoting
urllib.quote is used to quote urls so it considers '/' to be safe (as pointed
out in [1]). In here, we are quoting the image's name because it cannot have
spaces, but if a '/' appears, magento will throw a parse error to us, so,
it should not consider any character safe.

[1] http://docs.python.org/library/urllib.html#urllib.quote