Set LC_ALL too and now we have right decimal separators, monetary symbols, etc
locale.setlocale(locale.LC_ALL, lang) isn't really doing what we expected it
to do. Set LC_ALL to as this fixes a lot of problems for people who runs the
system in one locale and chooses Stoq to run in another.
Also, remove LANG as that doesn't seem to affect anything.