Correct the way PaymentGroup sums payments
When summing payments' total/paid_total/interest/penalty/discount/etc,
sometimes not all payments will be the same type (a sale for example can have
an outpayment because of a devolution). So, do that in the right way, that is,
SUM(inpayments) - SUM(outpayments) for sales/renegotiations and
SUM(outpayments) - SUM(inpayments) for purchases.

r=romaia