Pass total_amount_abs + paid_total to MultipleMethodSlave
When partially returning a partially paid sale, since there's a paid payment,
MultipleMethodSlave got confused and was reducing that paid payment from the
total (in that specific case, making the sum of paid payments being reduced
twice, since the total already had it reduced). Pass that paid_total together
so the calculation is always right.

r=romaia