Remove 'penalty' and 'discount' from returned sales
Make it simpler to do a return. Normally the person who is doing the return
doesn't have permissions to apply a discount/penalty. If that needs to be done,
it can be done on the payment itself when paying it.
Also, make quantities be formatted with format_quantity and improve the info
message that is displayed when a sale is returned and the client's debt changed

r=romaia