Fix 'get_all_classes' and correct some tests that weren't running
Something changed there and packages where being get with a leading '.' (this
didn't happen in the past probably, as the tests used to work). Workaround this
problem and correct some broken tests that were masked because of this
problem ('get_all_classes' where returning an empty list).

Also, remove 'test_editors' as we have now a more robust way of doing that.

r=romaia