Reenable some Domain tests and make some base editors private
Some baseeditors doesn't need to be public, as they are just being used on the
module itself. Also, all Domain tests are working (no need to skip none).
Add some more example data (and point some that already existed) so testDomain
can test as much domains as it can. There's still 2 skips because of the lack
of a PaymentRenegotiation and ProductIcmsTemplate examples.

r=romaia