Implement a wizard for trading products
POS now have a new menu, File -> New -> Trade. Running it will start a wizard
(based on the return wizard) for doing the trade, and the value returned is
used as a payment for the new sale on pos. Because of that, a new payment
method was created, 'trade'. It's hidden for now as it isn't intended to be
changed at the moment.
There's a new parameter that allows trade products from unknown sales. It is
false by default and it's very explicit in it's description that one should
only use it if he really understand what he is doing and at his own risk.

r=romaia