Avoid showing a message for missing dependencies
If a dependency is missing on jenkins (like poppler), if a message is show,
jenkins will hang. Avoid that by mocking DependencyChecker._error and also
put a return after those checks that try to use the failed to import module.
It's not a problem for normal stoq use since _error would raise SystemExit.
Also, simplify a little the mocks as this tests requires a lot of them. Makes
it easier to understand and also it's guaranted now that they will be unmocked
may the test success or fail.

r=romaia