Correct some 'testShow' and 'testCreate' that were being used wrong
We should always use 'testShow' on tests that runs a editor passing a model
and 'testCreate' on tests passing model=None.
Also, we should try to have tests for a editors defined on a module on the same
test module. eg: The module is "xxx.py" and it defines "X" and "Y" editors. We
should have a "test_xxx.py" module with "TestX" and "TestY" defined. This is
not easy to do right now, but when editing a test, we should check if we could
merge it with another one of the same module.

r=romaia