Mimic postgres rollback to savepoint on our objects
Consider the following savepoints:
    [A, B, C, D]
Rolling back to savepoint A is practically the same as rolling back to D, then
C, then B, then A. On our implementation of modified/created/deleted objects,
those sets got popped, so, the objects on savepoint D were discarded when in
this case, all of them should. The same for the savepoints, since we were only
removing, B, C and D would still be there. Also, consider the follwing:
    [A, B, C, D, A]
Postgres allow to rollback to A, them rollback to B, C or D and rollback to A
again (that would be the first A). On our implementation, we would only store a
savepoint name if it wasn't stored yet.

This will try to mimic postgres behaviour as much as possible. When rolling
back, we will remove modified/created/deleted sets of objects for each
savepoint rolled, and iterating in a reversed way will provide us that. Also,
it's now possible to do a savepoint A, then B, then A again, rollback to B and
rollback to first A, like the example above.

r=romaia