Workaround a problem when running selectBy passing a parent's class column as a keyword
Consider the following:
class _Base(Domain):
    x = ForeignKey('SomeOtherClass')

Class RealDomain(_Base):
    pass

Trying to RealDomain.selectBy(x=None) will produce an error. That because x
will be expanded to '_base.x_id' instead of 'real_domain.x_id'. It is a
problem specific to storm compatibility layer and should not exist
anymore when we remove that layer.
This commit works around that issue by 'redefining' the ForeignKey on
subclasses (like they defined the colum duplicated) so, it will be expanded
right.

r=romaia