When there's an empty space on header, it's considered a pseudo column
This is very strange. When saving columns, if there's an empty space at the
end, it will be returned as a column. That one doesn't have 'attribute' setted
since it was set by kiwi on the defined columns.
Workaround that problem and add an expand on PurchaseItemStep (make it more
elgant =P).

r=romaia