Do a savepoint/rollback on more run_dialogs done on same transaction
When we call run_dialog to an editor on the same transaction as the actual one,
if we don't rollback the changes done on it when it's cancelled, confirming
the actual editor would create phantom objects. And worse, since we can click
cancel even if mandatory data is missing, it would create inconsistent data, or
produce an error if the database has a restriction to it (not null for
instance).

r=romaia