Remove date from description of tillentries
There's no need to hardcode the date on the description, since TillEntry
already has a datetime column and it's show on the search. Also, the one on the
opening editor was contradictory, as the way it was written one would think %s
would be substituted by the value, not the date.

r=romaia