Remove pydispatch from external and use a simpler solution for events
pydispatch isn't maintained anymore, our fork is worse (since it don't get
updated). A solution to that would be using louie, that's a continuation of
pydispatch, but that would just increase our dependency list with something
that we barely use.
The solution here is very simple and easy to maintain. Also, it's using
weakrefs. Even that we don't use them on Stoq right now, it's prepared for it.
The Event compatibility was kept to work exactly as it was working with
pydispatch before, so, no api changes.

r=jdahlin