Allow connect same bound method from different objects on Event
There's an assert preventing the same callable from being connected to a event.
But, for a bound method, the function is the same, the difference is the object
that's passed as first argument. This now allows two objects from the same type
to connect the same bound method to the same event. Also, there's tests that
would've caught the problem.

r=romaia