Check if the transaction is closed before doing any operation
When trying to close an already closed transaction, an AttributeError would
happen on storm's side. Since it's not a very nice kind of error, we now check
if the transaction should already be closed and throw an InterfaceError in that
case.

Also, we check the transaction to see if it's closed on some operations
(commit, rollback, get, savepoint), since those should not be done on a closed
transaction.

r=romaia