Make it possible to close a loan on POS
It's possible now to close a loan on POS app. When doing it, instead of
creating a sale quote for sold items, those will be added on pos list like if
we added it by hand.

The loan's item step was improved to use 'SellableItemStep', just like the
trade does. That makes it easier and simpler.

Also, there's an improvement on current_transaction that makes it possible for
the trade/loan remain valid even after cancelling the checkout. This will
probably change when we support PAF-ECF, since it requires that if the checkout
fails, cancel everything.

r=romaia,jdahlin