Workaround to fix jenkins for a while
For some evil reason, jenkins is trying to import each module on plugins to
check for tests, and since they depend on the plugin.py file that adds them to
sys.path, they are raising import errors.
Evil reason because that is *only* happening on jenkins. I can't find a way to
reproduce it locally.

Doing this will make all the tests on plugins/to not run. This should be
reverted *ASAP* !!!!

r=jdahlin