Workaround for jenkins be able to test plugins and not fail
Nosetests try to search for tests in each .py file. Because of that, it was
raising ImportErrors on some plugins modules. For some strange reason, that
only happened on jenkins and not locally.
This workaround the problem by appending the plugins path on sys.path. We
should remove it when properly solving the plugins path issue.

By re-enabling the tests, a test on ecf was broken, but it's fixed now.

r=romaia,jdahlin