Add missing files after git migration
The tool we used to convert from bzr to git was not completely loseless.

This commits adds back the files that were missing or changes during
the migration. With this patch all the files in the the repository
have the same md5sum.

The tool used was git-remote-bzr with a couple of modifications,
rename support and work arounds for broken older bzr versions.