If sellable can be removed and closed, just show remove button
There are a lot of buttons on a sellable editor. "Ok", "Cancel", "Help" and
"Print Labels". If a sellable can be removed and closed, we would add 2 more
buttons for those actions. Instead, add just one, "Remove", since if it can be
closed/removed, the user will probably prefer to remove. Close is for sellables
that had stock one day and now haven't anymore.

r=romaia