If there are no objects for prefilling, return a fake one
The ProxyComboEntry's entry works by default with text mode. It will only
change to data mode when it gets prefilled with objects.
That's a problem specially with empty databases, e.g. when there is no
clients registered yet.
This is a workaround to avoid that problem and should be fixed on kiwi later
when we are not on RC period anymore. Then this code can be removed.

r=romaia,jdahlin