Abstract SellableItemStep to a slave
* Abstract SellableItemStep to be a slave, now called SellableItemSlave.
Sometime in the future we will move it to stoqlib.gui.slaves.sellableslave.
* Make SellableItemStep child of SellableItemSlave, so everything should be
working as it used to.
* Make all children of SellableItemStep call the parent's validate too, because
there's stuff done there.
* Remove _refresh_next api that was being called by children. Use the right api
for that.
* Implement an item_editor property that will be passed to AdditionSlave. This
simplifies the code for subclasses since they don't need to overwrite
post_init to do that, they just need to define it on the class.
* Fix a bug where unit_label were not being put on Settable model and because
of that it wasn't displaying the unit label.

Fixes: http://bugs.async.com.br/show_bug.cgi?id=5409

Change-Id: I454128e13f15d8090a6439717f85670907b5734a