Delivery: Cancelled dialog does not save changes.
If we edit a delivery and cancel the dialog, the changes are still
saved. If we edit the notes of a delivery and cancel the dialog, the
changes are also being saved.

The happens because the _CreateDeliveryModel is not a Domain and
therefore a store.rollback() has no effect and the rollback must be made
by hand.

Change-Id: I342d3dfb00f3cfab189c21d11f1c67c5bed0f5c7