Allow to reserve stock
Add a new column on SaleItem to check the quantity already decreased from
the stock. That way, when creating a SaleItem and indicating that some
quantity were already decreased, it won't be decreased again when selling.
Implement that for loans and work orders. For loan specially, that can prevent
a problem like that:
  1) The user closes a loan on sales app and it creates a sale of an item A
     with quantity of 10. Those 10 are returned to stock
  2) If someone sells that quantity, when trying to confirm the quote, it won't
     have enought quantity on the system for something the client didn't even
     return.
For work orders this means that we can decrease the stock as soon as the
product is used on that order.

Fixes http://bugs.async.com.br/show_bug.cgi?id=5412
Change-Id: Ic0323048621fe421e93264cf710caacb25c82ae4