CardOperarionCostEditor: fixes initialization.
CardOperationCostEditor would not initialize the upper limits of
installment start and end limits correctly (sometimes would not set
limits at all, or not allow limits at all).

It now uses the limits of the Card PaymentMethod to set the limits
correctly to the payment choices that allow installments and also does
not permit installments when it should not be allowed.

Change-Id: I1b0a8beced7b28d7b93b22019c6e40a4b365a3df