Add a report for workorder's quote
When we have a quote, allow the user to print a report for it.
It has a space to be signed by the responsible, so the client can take
home and use it as proof of what will be done with his equipment, at
what estimated cost and what estimated time.

Change-Id: I09e693bc8ca0a509b97d2c1c167d1c37406a05e3