Verifies stock right before changing it.
Sale, Transfer, Loan and StockDecrease:

All these stock operations now verify the availability of items in stock
right before changing it. This way, if these two operations start
simultaneously and one of them depletes the stock before the other, the
other operation will not happen. Instead, will warn the user that there
are missing items in the tried operation and show such items.

Change-Id: I29afb7168cfeb9c3abc030cd3a24a3194d107dc0