Fix a problem when selling the same value traded
This is a regression from commit 81750ea0.

When buying the same amount that was traded, it was generating a traceback when
the wizard wanted to create a payment with value 0.

When the total traded is equal to the sale's total, it should not be possible
to create more payments. The problem here was that we should use
get_total_to_pay instead of get_total_amount. That's true for the
store_credit's step.

Change-Id: If471afc23494f1a4f44ded742a6690e08ae1ba2f