Correct some wrong usage of yesno
When someone cancels the yesno dialog pressing esc, it's return value will be
False (just like if the user clicked on the "no" option).

Because of that, some inverse logic (where one would put the "yes" option in
the place where "no" belongs) could cause problems: e.g.: Someone pressing esc
when asked if he "Wants to cancel an order" and  the order being cancelled
anyway (just like if he clicked "yes")

So, since esc returns False, we should always be prepared to *don't do* the
action on False result.

Change-Id: Ia9208dc013e7d39bd6a385a18149418b5966bbc8