Explain a confusion when confirming a payment
When confirming a payment on payable/receivable, there's a confusion that the
payment will generate a till entry. That's not true.

So, if there's any money payment on it, show an infobar to alert the user that
the confirmation there won't generate a till entry so, if he needs to generate
one, he should do it by hand (cash addition/removal).

Also, add an explanation about that confusion on the manual and add
help_sections on them so if they have any confusion, they can understand why.

Change-Id: Iffce522442ec63774b5a58408c876c33a9c1f75d