Allow to create payments on quotes
The SaleQuoteWizard now has one more step: It allows to create payments
for the quote.

Those payments will be marked as pending so the user can mark them as paid
later (if they are really paid, obviously).

The SaleConfirmWizard is already prepared to take those payments in
consideration, so things should work out of the box there.

Change-Id: I1f208baf14d5a57353c7388877de4b0782b678f1