Remove STATUS_[UNAVAILABLE|BLOCKED] from Sellable
Sellable would change automatically from AVAILABLE/UNAVAILABLE depending on the
stock status (if we have stock or not). That can lead to lots of problems
(specially with synchronized bases), so remove it. If one needs to check the
stock availability, he should check the Storable.

STATUS_BLOCKED was removed together since it was never implemented. We should
implement the block feature as a flag in the future.

Some other changes were done because of that:
* POS: It was broken when selecting a sellable from the search (instead of
  getting from the barcode directly) where it wasn't checking the stock. Now
  everytime a product without stock enough is selected, a message will be
  displayed informing the user.
* Product: The 'Sellable.can_close' check now asks the product, that will see
  if it still has stock available (it's exactly what it did before by checking
  if it was unavailable)
* Sellable: Some sellable queries were not used anymore, so they were removed.
* SellableEditor: The unavailable_label was removed since it doesn't make sense
  anymore. It's not possible to change a service status by hand anymore so the
  user should use the 'Close' button that will accomplish the same.

Change-Id: I33d38d93f597c226d0d36fde32a8e942f48f789f