Correct get_available_sellables_query
get_available_sellables_query was wrong because it was removing the sellable
with id equals to the delivery service (it appears ok on tests and etc because
on most cases, they are the same since the delivery service is the first
service/sellable to be created and normally no one changes the parameter to
another sellable).

Also, kill get_available_sellables_for_quote_query that was doing the same as
get_available_sellables_query (but the right way) and use the second one on
it's only callsite and add a test for it.

Change-Id: Ib0957a6e1e194e43bdf9d6eecc082d360154f7d5