BaseEditor: change visual mode for GtkEntry
Instead of making them insensitive, mark them as not editable,
this allows the user to copy/paste information from the entries.

Probably needs to do the same for GtkTextView and combos and friends.

One change with this patch is that the visual mode is more subtle, it might
make sense to add another kind of indication that you cannot edit it,
like an infobar with a message, or changing the button from ok to close.

Change-Id: I95b46e486da8baf7f73cd7d2a3cd2addf8bb4c75