Avoid a false "sellable can't be sold" error
A corner case that can happen when there's some quantity already decreased, but
when calling "SaleItem.sell", the actual quantity on stock were 0 (and because
of that, the Storable would have set the sellable status to UNAVAILABLE), a
SellError could be raised.

This was resolved on master by removing the UNAVAILABLE status and doing some
modifications on that logic. In here, just don't check it for products (an
error will happen the same way if there's really no stock to decrease) to avoid
that problem that can happen a lot on WorkOrders (if all stock of a product
were used on the execution).

Change-Id: Ib0565529e1a27aa1b9006375a8c63b628787c775