Migrate quantity_decreased for renegotiated sales
The patch-04-12 didn't migrate renegotiated sales. So, duplicate the patch and
run it for renegotiated sales.

It's not a great problem since, after being renegotiated, there's nothing done
on the sale that would use the quantity_decreased information. But one day it
might be used, so lets have the data right and avoid future headaches.

Change-Id: Ia64228b3388182e0c0fc595e5b19bc015f9a9567
Fixes: http://bugs.async.com.br/show_bug.cgi?id=5455