Get the plugin name from .plugin file instead of dirname
Since now we can treat plugins as a package, it makes sense for us to make them
a module not named by the plugins' name

For example, we want to develop plugin 'foo', but we don't want our plugin to
conflict with another package 'foo'. Something like:

    stoqfoo
      __init__.py
      foo.plugin

Before, the plugin would be called 'stoqfoo' and so we would have to call the
.plugin file 'stoqfoo' and the name inside the plugin class (the one that gets
registered) 'stoqfoo' too. Now we can call the module whatever we want and have
the .plugin and the name on the class by the name we want.

At first it doesn't seem to make any difference. But we want to activate the
plugin by 'stoqdbadmin enable_plugin foo' (and not stoqfoo) and also, if the
plugin provides stoqdbadmin commands, we want to call them using
'stoqdbadmin foo <cmd>' instead of 'stoqdbadmin stoqfoo <cmd>'

For the existing plugins, since their module names and their .plugin have the
same name, nothing will change (but we can in the future if we want to).

Change-Id: I254c6a42841b25d6d0d32172db1a83b5a2058131