Add a way for the user to find Confirmed and Paid sales.
There are two statuses for valid sales: Confirmed (the sale was
confirmed but not all payments are paid yet) and Paid (the sale was
confirmed and all payments are paid).

So, when the user wants to see all valid sales (whether all payments are
paid or not), we need a special filter.

Also make the default search in the sales app search using the identifier as
well.

Change-Id: I201c4711183e449eb1e89d59b3c0bdb736a1182a