Make it easy to edit the payments on the checkout.
If the user choses the multiple method option, it is not easy to edit the
payments since we cant allow him to do that when using the TEF plugin (that
is required by the homologation).

If the user is not using the TEF plugin, let him edit the added payments,
and also, dont confirm the dialog when the full value is reached.

Change-Id: I8860f5a4d3dd148566ddfd02d47b7dd140285675
Fixes: http://bugs.async.com.br/show_bug.cgi?id=5499