Indicate the batches when confirming a sale
When we create a sale quote, it makes no sense to point the items' batches
since at that time, we don't really know (and are not interested) on the batch.
This can happen on an online sale too (like on magento).

For that reason, add a step that will set the batches properly for us.

Change-Id: I5292ddc388ec07de3eb853bf64497d9a921d4fe1