Log work order events
Create a new table, WorkorderHistory, that will hold information about
everything that has happened to the workorder.

Also, use that on WorkOrderHistorySlave instead of our actual workaround. Use
the NoteEditor to add notes on the history and remove the work order
package item's notes column since it's not needed anymore.

Change-Id: I8febd6789e2defdb55f50141ceb2c35f7ac94c4b