MultipleMethodStep: automatically fill value.
When the user selects a credit method (credit or store credit), fills
the payment value automatically with a correct value for such method.
Else, fills with the remaining value to be paid.

When updating the value because of a change of method, there's no need
to validate, instead, we should fill the value with a valid value if the
current one is invalid.

Change-Id: I04f65fe2bc619df5776fa510e5922db612e4b635