Convert id columns to use UUID
Change datatype for ID of all tables that has a te_id reference.
This will make it easier to syncronize objects between different
databases.

The following tables are not changed as they do not have a te_id
reference. Eg, they inherit from ORMObject, not Domain:

 - TransactionEntry
 - BranchSyncronization
 - CityLocation
 - Event
 - SystemTable

This requires the extension uuid-ossp to be installed, which
we verify before creating a new database.

Change-Id: I0253970bc125b4fbc5c8f13760e54d02272daf0c