Adds manager discount to price entries.
Every sellable as well as every user has a maximum discount they can
apply to a sellable. If the manager wants to allow a bigger discount for
a specific sale, they can click the error icon and open a dialog where
they can enter their credentials and authorize the discount.

Change-Id: I08feb357ede214fd06437d2c3fbb6b1c477f7328