Remove not oficial plugins from test
Since the commit 9295671c0fcec3cf91d78041cd80aa6823b47465, Stoq will
register plugins outside this repository (as long as they are on the same
checkout dir). Because of that, anyone having a plugin there would have a false
positive fail when testing pluginmanagerdialog. Avoid doing that and only list
the oficial plugins.

Also, check if the oficial plugins are *really* on the list.

Change-Id: Iab5bd4c1e6f07e58bcfb4b5b52c3b0f60210b780