Products without storable.
This will allow the users to create products that dont require stock
managment. This means that he will be able to decrease the stock (sell,
lend, etc...) Even if there are no purchases for the product.

Note that transfers and stock decreases are not allowed for products that
dont have storables (since it doesn't make sense).

Fixes: http://bugs.async.com.br/show_bug.cgi?id=3901
Change-Id: Ic88897c5f4563960c8f35d0c94a3343369262628