Optical quote: create payments when confirming
When trying to confirm a sale quote that has work orders associated with it
and these work orders are not finished, dont confirm the sale. Instead, just
create the payments for it.

A sale that has workorders can only be confirmed (ie, have the coupon
printed), when all the workorders are finished.

Change-Id: Ifc6ace54aa266a79fa5b0407126cbb4f885b5b5e