Duplicate items on abstractwizard
Instead of updating them, do a duplication. That way the user can change the
price for a set some items without having to do that for all and things like
that.

Because of that, add a new api on BatchDecreaseSelectionDialog to make batch
validation without appending them as the original ones.

Change-Id: I3317f304660345e8fff513882d31cf63dc679b5b