Allow to finish work orders without items
In some situations, it makes sense to finish an work order without any items.
e.g. Something was sent to warrant but it didn't need repair. In this case, we
didn't use any item to fix it and we won't charge the client for anything.

So, allow to finish work orders without any items and, for that specific case,
allow them to be closed on maintenance app directly (since it wont generate a
sale).

Change-Id: Ibbb1c0fa327b1d8071e44e78d8e623ddd03f8584