Show available batches on batch selection dialog
Add a list on batch selection dialog to show the available batches ordered by
their creation date (so the user can choose the oldest batch).

When clicking on a batch on that list, that batch will be appended (or have
it's quantity's focus grabed if it was already there) as a helper to add
quantity for that batch.

Change-Id: I28f3e3f77f5000a345f3392b8e0566cdefe040d1